Update MainPage.xaml.cs
- BISC_toBindを別ファイルに分離
- dllをうまく見つけてくれるようになったため, 解決のための実装を削除
- ディレクトリ指定関連をメソッド化
- WidthがNaNであることに備え, ActualWidthを採用することにした
- ほか, バグ修正
